:recycle: refactor : security 설정 위해 controller 분리 및 엔드포인트 수정
