+++ src/main/java/umc/unimade/domain/qna/controller/QnAController.java
@@ -10,24 +10,19 @@ import org.springframework.web.bind.annotation.*;
 import umc.unimade.domain.accounts.entity.Buyer;
 import umc.unimade.domain.accounts.entity.Seller;
 import umc.unimade.domain.qna.dto.AnswerCreateRequest;
-import umc.unimade.domain.qna.dto.AnswerResponse;
 import umc.unimade.domain.qna.dto.QuestionCreateRequest;
-import umc.unimade.domain.qna.dto.QuestionResponse;
 import umc.unimade.domain.qna.service.QnACommandService;
-import umc.unimade.domain.qna.service.QnAQueryService;
 import umc.unimade.global.common.ApiResponse;
 import umc.unimade.global.common.ErrorCode;
 import umc.unimade.domain.accounts.exception.UserExceptionHandler;
 import umc.unimade.global.security.LoginBuyer;
 import umc.unimade.global.security.LoginSeller;
 
-//TO DO : Buyer인지 Seller인지 Role 검증
 @RestController
-@RequestMapping("/api/qna")
+@RequestMapping("/qna")
 @RequiredArgsConstructor
 public class QnAController {
     private final QnACommandService qnaCommandService;
-    private final QnAQueryService qnaQueryService;
 
     @Tag(name = "QnA", description = "qna 관련 API")
     @Operation(summary = "질문 작성")
@@ -45,12 +40,15 @@ public class QnAController {
         }
     }
 
+
     @Tag(name = "QnA", description = "qna 관련 API")
-    @Operation(summary = "질문 불러오기")
-    @GetMapping("/question/{questionId}")
-    public ResponseEntity<ApiResponse<QuestionResponse>> getQuestion(@PathVariable Long questionId) {
+    @Operation(summary = "질문 삭제")
+    @DeleteMapping(value = "/question/{questionId}")
+    public ResponseEntity<ApiResponse<Void>> deleteQuestion(@PathVariable Long questionId,
+                                                            @LoginBuyer Buyer buyer) {
         try {
-            return ResponseEntity.ok(ApiResponse.onSuccess(qnaQueryService.getQuestion(questionId)));
+            qnaCommandService.deleteQuestion(questionId, buyer);
+            return ResponseEntity.ok(ApiResponse.onSuccess(null));
         } catch (IllegalArgumentException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ApiResponse.onFailure(HttpStatus.BAD_REQUEST.name(), e.getMessage()));
         } catch (UserExceptionHandler e) {
@@ -71,35 +69,6 @@ public class QnAController {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ApiResponse.onFailure(HttpStatus.BAD_REQUEST.name(), e.getMessage()));
         } catch (UserExceptionHandler e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.onFailure(ErrorCode.SELLER_NOT_FOUND.getCode(), ErrorCode.SELLER_NOT_FOUND.getMessage()));
-        }
-    }
-
-    @Tag(name = "QnA", description = "qna 관련 API")
-    @Operation(summary = "답변 불러오기")
-    @GetMapping("/answer/{answerId}")
-    //To do : user 토큰 추가
-    public ResponseEntity<ApiResponse<AnswerResponse>> getAnswer(@PathVariable Long answerId) {
-        try {
-            return ResponseEntity.ok(ApiResponse.onSuccess(qnaQueryService.getAnswer(answerId)));
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ApiResponse.onFailure(HttpStatus.BAD_REQUEST.name(), e.getMessage()));
-        } catch (UserExceptionHandler e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.onFailure(ErrorCode.SELLER_NOT_FOUND.getCode(), ErrorCode.SELLER_NOT_FOUND.getMessage()));
-        }
-    }
-
-    @Tag(name = "QnA", description = "qna 관련 API")
-    @Operation(summary = "질문 삭제")
-    @DeleteMapping(value = "/question/{questionId}")
-    public ResponseEntity<ApiResponse<Void>> deleteQuestion(@PathVariable Long questionId,
-                                                            @LoginBuyer Buyer buyer) {
-        try {
-            qnaCommandService.deleteQuestion(questionId, buyer);
-            return ResponseEntity.ok(ApiResponse.onSuccess(null));
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ApiResponse.onFailure(HttpStatus.BAD_REQUEST.name(), e.getMessage()));
-        } catch (UserExceptionHandler e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.onFailure(ErrorCode.BUYER_NOT_FOUND.getCode(), ErrorCode.BUYER_NOT_FOUND.getMessage()));
         }
     }
